requirements.txt
django
graphene-django

djangorestframework
markdown
django-filter

celery

requests

gevent
gunicorn
